requirements.txt
gymnasium==1.0.0
langchain_openai==0.3.7
matplotlib==3.10.1
networkx==3.3
Pillow==9.5.0
Pillow==11.1.0
plotly==5.18.0
PyYAML==6.0.1
PyYAML==6.0.2
Requests==2.32.3
setuptools==69.5.1
